Discovered and fixed bugs in CLI
I discovered that if you try adding a field or method to a non-existing class it will display that it did but in console show it did not. To fix this I added a classExist function in UMLClass so I could do a boolean check and see if the class exists already or not. I also added a list AllClassInfo command so you can see all the classes you made with their fields and methods.