Added Autocomplete Functionality in Swing
Added Autocomplete with a set list of commands that when pressed Tab will fill out the command partially.

Look to the right of the text box and in blue the preview of the autocompleted word will be there.

Also decoupled the original CLI class so its easier to read and modify.